Fl_Config read/write methhods re-written. This update deletes all old read_*/write_* methods. Why? Cause they not safe, And they allocated/de-allocated memory all the time, which is not good. So, replacement: read_* -> read(key, ret_value, def_value) write_* -> write(key, value) These functions assumes that current section set by set_section. Or functions without need for set_section: set(section, key, ret_value, def_value) get(section, key, value)
All methods returns error now, so it's possible to do:
if(!config.read(...)) { do something }

It is also possible to pass non-allocated char pointer to
read functions, it needs to be freed by user then. Allocation
is possible with pointer-to-reference system.

Sorry, about changing this again, but speed and small
memory footprint is our trademark! :)

----------------

Added Fl_Config(vendor, application, mode)
This could easily used with apps. mode is either USER or SYSTEM
if USER then filename is
$(home)/.ede/apps/$(application)/$(application).conf
or if SYSTEM it's
$(prefix)/share/ede/apps/$(application)/$(application).conf

-------------------