Added Fl_Menu_() contructor, cause update from Bill removed end() function from Fl_Menu_(...), That was cause Fl_Item_Group is now inherited from Fl_Menu_ and thats Groups normal behaviour.
But... It caused problems, cause we couple of static menus, like in
Fl_Text_Editor, that set Fl_Group::current() before any window was
created. So this broke most of the programs...

Fl_Menu_() contructs 0x0 menu, and does end() there. So it can be used as
compiled-in static menu.