Fixed some serious desing mistakes..
 - Added styles
 - copy/set children styles in draw()
 - DO NOT use layout_align() feature in widget!!
 - Made it resizable also other way than preferred size..
 - Do not call child->resize after Fl_Widget::layout, this will cause
layout called recursively..

Why not to use layout_align() in widgets?
Morelikely it leads to widget, which is not resizable
in all cases. i.e. Im not using layout_align() stuff, and set my
Fl_Date_Input height to 40 and width to 300, It must look still correct!
It also leads to slowness, it is stupid to make simple task slower than it
can be.. Although "so called" low-level (alex)? stuff can and IMHO should
be in core widgets.. It's totally different case when building an
application!

Compare:
layout() {
  do some stuff;
  Fl_Group::layout();
}

AND

layout() {
  do some stuff;
  child1->resize
  child2->resize
  Fl_Widget::layout()
}

This is example of widget that has 2 children (Fl_Date_Input).
later layout() will result much faster and convient layout..